+++ src/main/java/com/smoorbuilderserver/repository/ClassDescriptionRepository.java
@@ -1,6 +1,5 @@
 package com.smoorbuilderserver.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -9,11 +8,11 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.ClassDescription;
 
 @Repository
-public interface ClassDescriptionRepository extends CrudRepository<ClassDescription, BigInteger> {
+public interface ClassDescriptionRepository extends CrudRepository<ClassDescription, Long> {
 	
 	public ClassDescription findByClassName(String generalClassName);
 	
-	public ClassDescription findByid(BigInteger id);
+	public ClassDescription findByid(Long id);
 	
 	public List<ClassDescription> findDistinctClassDescriptionByClassName(String generalClassName);
 }
