+++ src/main/java/com/smoorbuilderserver/repository/ProductionSkillDescriptionRepository.java
@@ -1,6 +1,5 @@
 package com.smoorbuilderserver.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -9,11 +8,11 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.ProductionSkillDescription;
 
 @Repository
-public interface ProductionSkillDescriptionRepository extends CrudRepository<ProductionSkillDescription, BigInteger> {
+public interface ProductionSkillDescriptionRepository extends CrudRepository<ProductionSkillDescription, Long> {
 	
 	public ProductionSkillDescription findByProductionSkillName(String productionlName);
 	
-	public ProductionSkillDescription findByid(BigInteger id);
+	public ProductionSkillDescription findByid(Long id);
 	
 	public List<ProductionSkillDescription> findDistinctProductionSkillDescriptionByProductionSkillName(String productionName);
 }
