+++ src/main/java/com/smoorbuilderserver/repository/RaceDescriptionRepository.java
@@ -1,6 +1,5 @@
 package com.smoorbuilderserver.repository;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
@@ -9,11 +8,11 @@ import org.springframework.stereotype.Repository;
 import com.smoorbuilderserver.model.RaceDescription;
 
 @Repository
-public interface RaceDescriptionRepository extends CrudRepository<RaceDescription, BigInteger> {
+public interface RaceDescriptionRepository extends CrudRepository<RaceDescription, Long> {
 	
 	public RaceDescription findByRaceName(String raceName);
 	
-	public RaceDescription findByid(BigInteger id);
+	public RaceDescription findByid(Long id);
 	
 	public List<RaceDescription> findDistinctRaceDescriptionByRaceName(String raceName);
 }
